Change update-games to new design
